Fixed problem with size of activity message. If it's too big, the font size is modified. Also cut words too long.

+++ src/ClicActivity/Grid.py
@@ -109,7 +109,7 @@ class Grid(object):
                 
             except:
                 self.cellHeight =  float(30)
-                self.cellWidth =   float(30)
+                self.cellWidth =   float(40)
 
             self.hasBorder = False
             try:
